Update project1-libc-redo.c
This prints " ` " to the console the first time you run it, and then it reads in the user input one character at a time. Hit "Run," it will read one and then pause. Hit "Run" again, it will read the next, and so on until you keep hitting run for it to read in your entire input. Once you hit "Run" again after reading everything in, it will finally print an "r" to the console before stopping.